requirements.txt
numpy
torch
diffusers
pillow
scikit-image
rasterio
matplotlib